+++ index.html
@@ -7,7 +7,6 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кексобукинг</title>
 
-  <script src="js/main.js"></script>
 </head>
 <body>
 
@@ -230,5 +229,6 @@
     </div>
   </template>
 
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
